+++ front-end/src/store/adapters/adaptersToDto.ts
@@ -1,6 +1,8 @@
 import { CreateOfferDto, LocationDto } from '../dtos';
 import {Location, NewComment, NewOffer} from '../../types/types';
 import { CreateCommentDto} from '../dtos/commments/index.js';
+import {MIN_OFFER_IMAGES, OFFER_STATIC_URLS} from '../../const';
+import {getRandomItem, getSlicedRandomArray} from '../../utils';
 
 const locationToDto = (dto: Location): LocationDto => ({lat: dto.latitude, long: dto.longitude});
 export const offerToDto = (offer: NewOffer): CreateOfferDto => ({
@@ -14,7 +16,9 @@ export const offerToDto = (offer: NewOffer): CreateOfferDto => ({
   postDate: new Date().toISOString(),
   isPremium: offer.isPremium,
   city: offer.city.name,
-  location: locationToDto(offer.location)
+  location: locationToDto(offer.location),
+  previewUrl: getRandomItem(OFFER_STATIC_URLS),
+  urls: getSlicedRandomArray(OFFER_STATIC_URLS, MIN_OFFER_IMAGES)
 });
 
 export const commentToDto = (comment: NewComment, offerId: string): CreateCommentDto => ({
