+++ src/shared/modules/offer/index.ts
@@ -7,5 +7,4 @@ export * from './types/index.js';
 export * from './offer-service.pipelines.js';
 export * from './offer.container.js';
 export * from './offer.controller.js';
-export * from './offer.constants.js';
 export * from './offer.messages.js';
