--- src/shared/modules/offer/offer.constants.ts
@@ -1,9 +1,0 @@
-export const OFFER_STATIC_URLS = [
-  'amsterdam.jpg',
-  'amsterdam@2x.jpg',
-  'apartment-01.jpg',
-  'apartment-02.jpg',
-  'apartment-03.jpg',
-  'apartment-small-03.jpg',
-  'apartment-small-04.jpg',
-];
